Added save button in photo details
