+++ Electrolux-task/View/ItemsTab/ItemCell.swift
